requirements.txt
numpy
Pillow
scipy
matplotlib
opencv-python
